Reorganize file code into binary and json directories
