+++ include/BLIB/Files.hpp
@@ -6,9 +6,8 @@
  * @brief Collection of classes for dealing with different data file types
  */
 
-#include <BLIB/Files/BinaryFile.hpp>
+#include <BLIB/Files/Binary.hpp>
 #include <BLIB/Files/FileUtil.hpp>
 #include <BLIB/Files/JSON.hpp>
-#include <BLIB/Files/Schema.hpp>
 
 #endif
