+++ include/BLIB/Files/Binary.hpp
@@ -1,0 +1,6 @@
+#ifndef BLIB_FILES_BINARY_HPP
+#define BLIB_FILES_BINARY_HPP
+
+#include <BLIB/Files/Binary/BinaryFile.hpp>
+
+#endif
