+++ include/BLIB/Files/Binary/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(BLIB PUBLIC
+    BinaryFile.hpp
+)
