+++ include/BLIB/Files/JSON/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(BLIB PUBLIC
+    JSON.hpp
+    Schema.hpp
+)
