+++ src/Files/CMakeLists.txt
@@ -1,7 +1,5 @@
 target_sources(BLIB PRIVATE
     FileUtil.cpp
-    JSON.cpp
-    JSONLoader.hpp
-    JSONLoader.cpp
-    Schema.cpp
 )
+
+add_subdirectory(JSON)
