+++ src/Files/JSON/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(BLIB PRIVATE
+    JSON.cpp
+    JSONLoader.hpp
+    JSONLoader.cpp
+    Schema.cpp
+)
