+++ src/Files/JSON/Schema.cpp
@@ -1,4 +1,4 @@
-#include <BLIB/Files/Schema.hpp>
+#include <BLIB/Files/JSON/Schema.hpp>
 
 #include <BLIB/Logging.hpp>
 #include <algorithm>
