+++ tests/Files/Binary/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(BLIB.t PUBLIC
+    BinaryFile.t.cpp
+)
