+++ tests/Files/JSON/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(BLIB.t PUBLIC
+    JSON.t.cpp
+    Schema.t.cpp
+)
